+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v0.7.0 (2020-08-01)
 
 **The backup structure has changed! Read below for more detail.**
 
@@ -16,7 +16,10 @@
   * Backup structure is now human-readable.
   * App window now has a minimum size, 640x480.
     (Note: For now, the crates.io release will not have a minimum size.)
+  * File size units are now adjusted based on the size, rather than always using MiB.
+    ([contributed by wtjones](https://github.com/mtkennerly/ludusavi/pull/32))
 
+### New backup structure
 Previously, Ludusavi used Base64 to encode game names and original paths when
 organizing backups. There were some technical advantages of that approach, but
 it was not easy to understand, and there was a technical flaw because Base64